+++ apps/widget-configurator/src/app/configurator/utils/validateURL.ts
@@ -1,5 +1,0 @@
-import { parseENSAddress, uriToHttp } from '@cowprotocol/common-utils'
-
-export const validateURL = (url: string): boolean => {
-  return uriToHttp(url).length > 0 || Boolean(parseENSAddress(url))
-}