+++ libs/ui/src/containers/Footer/index.tsx
@@ -9,8 +9,8 @@ import SVG from 'react-inlinesvg'
 import {
   getGlobalFooterNavItems,
   GLOBAL_FOOTER_DESCRIPTION,
-  SOCIAL_LINKS,
   PRODUCT_LOGO_LINKS,
+  SOCIAL_LINKS,
 } from './footer.constants'
 import { FooterAnimation } from './footerAnimation'
 import {
@@ -36,6 +36,8 @@ import {
 import { UI } from '../../enum'
 import { MenuItem } from '../../pure/MenuBar'
 import { ProductLogo, ProductVariant } from '../../pure/ProductLogo'
+
+export { getGlobalFooterNavItems } from './footer.constants'
 
 export interface NavItemProps extends Omit<MenuItem, 'label' | 'badge'> {
   label?: string
@@ -93,7 +95,7 @@ const FooterLink = ({ href, external, label, utmSource: _utmSource, utmContent, 
   )
 }
 
-export const GLOBAL_FOOTER_NAV_ITEMS = getGlobalFooterNavItems(false)
+export const GLOBAL_FOOTER_NAV_ITEMS = getGlobalFooterNavItems()
 
 export const Footer = ({
   description = GLOBAL_FOOTER_DESCRIPTION,